+++ app/scripts/components/trend-diagram.js
@@ -654,7 +654,7 @@ angular.module("korpApp").component("trendDiagram", {
                         try {
                             abs_y = series.abs_data[i].y
                         } catch (e) {
-                            c.log("i", i, x)
+                            console.log("i", i, x)
                         }
 
                         const rel = series.name + ":&nbsp;" + formattedY
@@ -761,7 +761,7 @@ angular.module("korpApp").component("trendDiagram", {
                     $timeout(() => renderGraph(Rickshaw, graphData, cqp, labelMapping, currentZoom, showTotal))
                 } catch (e) {
                     $timeout(() => {
-                        c.error("graph crash", e)
+                        console.error("graph crash", e)
                         $ctrl.localUpdateLoading(false)
                         $ctrl.error = true
                     })