+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Replaced Raphael library with Chart.js, used in the pie chart over corpus distribution in statistics
+- Removed the global `c` alias for `console`
 
 ## [9.6.0] - 2024-05-27